Moved LisContextType to Lis\v1
